frontend/requirements.txt
flask
plotly
google-cloud-secret-manager
google-cloud-bigquery
pandas
db-dtypes